+++ src/app/api/auth/logout/route.ts
@@ -2,8 +2,9 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-import { authOptions, getToken } from '@/utils/auth';
 import { getServerSession } from 'next-auth';
+import { authOptions } from '../[...nextauth]/route';
+import { getToken } from '../auth';
 
 export async function GET() {
   const session = await getServerSession(authOptions);
