+++ src/services/discovery/datasetGet.ts
@@ -2,10 +2,10 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+import { ExtendedSession } from '@/app/api/auth/auth.types';
 import axios from 'axios';
 import { RetrievedDataset } from './types/dataset.types';
 import { createHeaders } from './utils';
-import { ExtendedSession } from '@/utils/auth';
 
 export const makeDatasetGet = (discoveryUrl: string) => {
   return async (id: string, session?: ExtendedSession): Promise<RetrievedDataset> => {
