+++ src/services/discovery/datasetList.ts
@@ -2,10 +2,10 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+import { ExtendedSession } from '@/app/api/auth/auth.types';
 import axios, { AxiosResponse } from 'axios';
-import { DatasetSearchQuery, DatasetSearchOptions, DatasetsSearchResponse } from './types/datasetSearch.types';
+import { DatasetSearchOptions, DatasetSearchQuery, DatasetsSearchResponse } from './types/datasetSearch.types';
 import { createHeaders } from './utils';
-import { ExtendedSession } from '@/utils/auth';
 
 export const makeDatasetList = (discoveryUrl: string) => {
   return async (options: DatasetSearchOptions, session?: ExtendedSession): Promise<AxiosResponse<DatasetsSearchResponse>> => {
